requirements.txt
scipion-em
scipion-pyworkflow